Move disable() to callable.py
